Reverted %a with original 32bit array length version. Added %A type for 64bit.
Reverted %a back to the original format string type so that it takes
a 32bit array length array. Moved 64bit versions to a new format
string type %A .

+++ src/ChildNode.C
@@ -269,7 +269,7 @@ int ChildNode::proc_SetTopoEnv( PacketPtr ipacket ) const
     char* sg_byte_array = NULL;
     int* keys = NULL;
     char** vals = NULL;
-    uint64_t i, count;
+    uint32_t i, count;
     NetworkTopology* nt = _network->get_NetworkTopology();    
 
     if( ipacket->unpack( "%s %ad %as", &sg_byte_array, 
@@ -665,18 +665,18 @@ int ChildNode::proc_PortUpdate( PacketPtr ipacket ) const
         
         if( _network->is_LocalNodeBackEnd() )
             s->send( PROT_TOPO_UPDATE, "%ad %aud %aud %as %auhd",
-                     &type, uint64_t(1),
-                     &send_iprank, uint64_t(1),
-                     &send_myrank, uint64_t(1),
-                     &host_arr, uint64_t(1),
-                     &send_port, uint64_t(1) );
+                     &type, 1,
+                     &send_iprank, 1,
+                     &send_myrank, 1,
+                     &host_arr, 1,
+                     &send_port, 1 );
         else
             s->send_internal( PROT_TOPO_UPDATE, "%ad %aud %aud %as %auhd",
-                              &type, uint64_t(1),
-                              &send_iprank, uint64_t(1),
-                              &send_myrank, uint64_t(1),
-                              &host_arr, uint64_t(1),
-                              &send_port, uint64_t(1) );
+                              &type, 1,
+                              &send_iprank, 1,
+                              &send_myrank, 1,
+                              &host_arr, 1,
+                              &send_port, 1 );
         s->flush();
         free(host_arr);
     }
