+++ src/ParentNode.C
@@ -445,7 +445,7 @@ Stream * ParentNode::proc_newStream( PacketPtr ipacket ) const
 {
     Stream* stream;
     Rank* backends = NULL;
-    uint64_t num_backends;
+    uint32_t num_backends;
     unsigned int stream_id;
     int tag, ds_filter_id, us_filter_id, sync_id;
 
@@ -593,7 +593,7 @@ int ParentNode::proc_deleteStream( PacketPtr ipacket ) const
 int ParentNode::proc_newFilter( PacketPtr ipacket ) const
 {
     int retval = 0;
-    uint64_t nfuncs = 0;
+    uint32_t nfuncs = 0;
     char* so_file = NULL;
     char** funcs = NULL;
     unsigned short* fids = NULL;
@@ -718,8 +718,8 @@ int ParentNode::proc_NewChildDataConnection( PacketPtr ipacket, int isock )
         char* topo_dup = strdup( topo_str.c_str() );
         PacketPtr pkt( new Packet(CTL_STRM_ID, PROT_NET_SETTINGS, "%s %ad %as", 
                                   topo_dup,
-                                  keys, uint64_t(count), 
-                                  vals, uint64_t(count)) );
+                                  keys, count, 
+                                  vals, count) );
         pkt->set_DestroyData( true );
         child_node->sendDirectly( pkt );
     }
@@ -765,11 +765,11 @@ int ParentNode::proc_NewChildDataConnection( PacketPtr ipacket, int isock )
                 Port dummy_port = UnknownPort;
                 char* dummy_host = strdup("NULL"); // ugh, this needs to be fixed
                 s->send_internal( PROT_TOPO_UPDATE, "%ad %aud %aud %as %auhd", 
-                                  &type, uint64_t(1), 
-                                  &my_rank, uint64_t(1), 
-                                  &child_rank, uint64_t(1), 
-                                  &dummy_host, uint64_t(1), 
-                                  &dummy_port, uint64_t(1) );
+                                  &type, 1, 
+                                  &my_rank, 1, 
+                                  &child_rank, 1, 
+                                  &dummy_host, 1, 
+                                  &dummy_port, 1 );
                 free( dummy_host );
             }
         }
